journal_api/requirements.txt
fastapi~=0.111.0
uvicorn[standard]~=0.29.0

# Database
sqlalchemy[asyncio]~=2.0.29
asyncpg~=0.29.0

# Config & Secrets
pydantic-settings~=2.2.0
python-dotenv>=1.0.0
google-cloud-secret-manager~=2.19.0

# Storage (optional for attachments)
google-cloud-storage>=2.17.0

# Auth (token handling if needed)
python-jose[cryptography]>=3.3.0

